+++ example-native/views/ui1.js
@@ -7,13 +7,13 @@ const template = /*html*/`
   <div class="field">
     <label for="" class="label">bwc-combobox - multiple</label>
     <div class="control has-icons-left">
-      <bwc-combobox ref="mcRef" :tags="mc.tags" multiple input-class="input" listid="multiple-ac" required :items="ac.items" v-model="ac.multipleValue" @search="(e) => autoComplete(e)" @selected="selected"></bwc-combobox>
+      <bwc-combobox ref="mcRef" :tags="mc.tags" tag-limit="3" multiple input-class="input" listid="multiple-ac" required :items="ac.items" v-model="ac.multipleValue" @search="(e) => autoComplete(e)" @select="selected"></bwc-combobox>
     </div>
   </div>
   <div class="field">
     <label for="" class="label">bwc-combobox - single</label>
     <div class="control has-icons-left">
-      <bwc-combobox listid="single-ac" required :items="ac.items" v-model="ac.singleValue" @search="(e) => autoComplete(e)" @selected="selected"></bwc-combobox>
+      <bwc-combobox listid="single-ac" required :items="ac.items" v-model="ac.singleValue" @search="(e) => autoComplete(e)" @select="selected"></bwc-combobox>
     </div>
   </div>
 
@@ -27,36 +27,16 @@ const template = /*html*/`
   <div class="field">
     <label for="" class="label">bwc-combobox - Search Country</label>
     <div class="control has-icons-left">
-      <bwc-combobox object-key="key" object-text="text" listid="country2" required :items="country.items" v-model="country.value" @search="(e) => countrySearch(e)" @selected="countrySelected"></bwc-combobox>
+      <bwc-combobox object-key="key" object-text="text" listid="country2" required :items="country.items" v-model="country.value" @search="(e) => countrySearch(e)" @select="countrySelected"></bwc-combobox>
     </div>
   </div>
 
   <div class="field">
     <label for="" class="label">bwc-combobox - Search State (dependent on country)</label>
     <div class="control has-icons-left">
-      <bwc-combobox multiple object-key="key" object-text="text" listid="state2" required :items="state.items" :tags="state.tags" @search="(e) => stateSearch(e)" @selected="stateSelected"></bwc-combobox>
+      <bwc-combobox multiple object-key="key" object-text="text" listid="state2" required :items="state.items" :tags="state.tags" @search="(e) => stateSearch(e)" @select="stateSelected"></bwc-combobox>
     </div>
   </div>
-
-  <hr/>
-
-  <h3>dependent autocomplete & object search example</h3>
-
-  <!-- TOREMOVE
-  <div class="field">
-    <div class="control">
-      <label for="" class="label">Search Country</label>
-      <bwc-autocomplete listid="country" required :items="country.items" v-model="country.value" @search="(e) => countrySearch(e)" @selected="countrySelected"></bwc-autocomplete>
-    </div>
-  </div>
-
-  <div class="field">
-    <div class="control">
-      <label for="" class="label">Search State</label>
-      <bwc-autocomplete listid="state" required :items="state.items" v-model="state.value" @search="(e) => stateSearch(e)" @selected="stateSelected"></bwc-autocomplete>
-    </div>
-  </div>
-  -->
 
   <hr/>
 
@@ -197,7 +177,7 @@ export default {
     onMounted(async () => {
       console.log('ui1 mounted!')
       console.log('mcRef', mcRef)
-      mc.tags = ['aa1', 'aa2', 'aa1']
+      mc.tags = ['aa1', 'aa2', 'aa1', 'aa3', 'aa4']
     })
 
     return {